POCImplementations/PoCAIPipeline/requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.30.0
peft>=0.5.0

# MLOps
mlflow>=2.8.0

# Feature Store
feast[redis]>=0.36.0
redis>=4.5.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# Image Processing
opencv-python>=4.8.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0

